% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_matrix.R
\name{ci_matrix}
\alias{ci_matrix}
\title{Create confidence intervals for all coefficients.}
\usage{
ci_matrix(model_list, model_names = NULL, levels = c(0.95, 0.5))
}
\arguments{
\item{model_list}{A list of regression models.}

\item{model_names}{A list of names for the regression models (default is
\code{NULL}).}

\item{levels}{At most two confidence levels.}
}
\value{
A data frame of model coefficients with their respective
 confidence intervals.
}
\description{
Create confidence intervals for all coefficients across all models, excluding
the intercept.
}
\details{
This function is called from \code{\link{coefficient_forestplot}} and might
not be of direct use for most users. Confidence intervals are obtained through
\code{confint} in the \code{stats} package. Intercepts are excluded from the
analysis as they don't relate to interactions or "effect modification," but
would be most analogous to an average effect.
}
\examples{
 # Simulate data

 N = 500

 m = rep(1:4, each=N)

 x1 = rnorm(n=N*4) x2 = rnorm(n=N*4) x3 = rnorm(n=N*4)

 y = x1 + x2 + x3 + rnorm(n=N*4)

 dat = data.frame(m, x1, x2, x3, y)

 m1 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==1)
 m2 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==2)
 m3 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==3)
 m4 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==4)

 mList = list(m1, m2, m3, m4)

 # Create data frame

 ci_matrix(model_list = mList, levels = 0.95)
}
