% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_anova.R
\name{coefficient_anova}
\alias{coefficient_anova}
\title{Compare coefficients of all shared predictors across models via ANOVA.}
\usage{
coefficient_anova(model_list, model_names = NULL,
  padj = p.adjust.methods)
}
\arguments{
\item{model_list}{A list of regression models.}

\item{model_names}{A list of names for the regression models (default is
\code{NULL}).}

\item{padj}{Adjustment of p-values for multiple comparisons. Must come from
\code{\link{p.adjust.methods}}.}
}
\value{
Data frame of coefficient information. Data frame includes coefficient
 estimates for each model as well as results from ANOVA test comparing
 coefficients across models.
}
\description{
Compare coefficients of all shared predictors across models via ANOVA using
\code{\link{coef_welch}}.
}
\details{
For each predictor variable in the models, runs \code{\link{coef_welch}} to
test for equality of regression parameters across models.
}
\examples{
 # Simulate data N = 500

 m = rep(1:4, each=N)

 x1 = rnorm(n=N*4)
 x2 = rnorm(n=N*4)
 x3 = rnorm(n=N*4)

 y = x1 + x2 + x3 + rnorm(n=N*4)

 dat = data.frame(m, x1, x2, x3, y)

 m1 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==1)
 m2 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==2)
 m3 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==3)
 m4 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==4)

 mList = list(m1, m2, m3, m4)

 # Compare coefficients across models

 coefficient_anova(model_list = mList, model_names = c('Group A', 'Group B',
 'Group C', 'Group D'))

}
\seealso{
\code{\link{coef_welch}}
}
