% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_forestplot.R
\name{coefficient_forestplot}
\alias{coefficient_forestplot}
\title{Create forestplot of model coefficients with confidence intervals.}
\usage{
coefficient_forestplot(model_list, model_names = NULL, levels = c(0.95,
  0.5), horiz = TRUE)
}
\arguments{
\item{model_list}{A list of regression models.}

\item{model_names}{A list of names for the regression models (default is
\code{NULL}).}

\item{levels}{A list of levels of confidence for confidence intervals (max 2
levels).}

\item{horiz}{Toggle whether confidence intervals are displayed horizontally
or vertically. default is set to horizontal (\code{horiz = TRUE}).}
}
\value{
A ggplot object to compare model coefficient estimates with their
  corresponding confidence interval(s), grouped by coefficient.
}
\description{
Create a ggplot forestplot of model coefficients with confidence intervals.
}
\details{
The forest plot orders variables along the axis and colors the data by model.
If \code{model_names = NULL}, the default, models are numbered sequentially
in the order they appear in \code{model_list} (Model 1, Model 2, Model 3,
etc.). If two confidence levels are provided the shorter interval is drawn
with a thicker line.
}
\examples{
 # Simulate data N = 500

 m = rep(1:4, each=N)

 x1 = rnorm(n=N*4)
 x2 = rnorm(n=N*4)
 x3 = rnorm(n=N*4)

 y = x1 + x2 + x3 + rnorm(n=N*4)

 dat = data.frame(m, x1, x2, x3, y)

 m1 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==1)
 m2 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==2)
 m3 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==3)
 m4 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==4)

 mList = list(m1, m2, m3, m4)

 # Create plot

 coefficient_forestplot(model_list = mList,
     model_names =c('Group A', 'Group B', 'Group C', 'Group D'),
     horiz = FALSE)

}
