% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resid_df.R
\name{get_resid_df}
\alias{get_resid_df}
\title{Extract residuals from multiple models.}
\usage{
get_resid_df(model_list, model_names = NULL)
}
\arguments{
\item{model_list}{A list of regression models.}

\item{model_names}{A list of names for the regression models (default is
\code{NULL}).}
}
\value{
A data frame of residuals and their corresponding model names.
}
\description{
Extracts residuals from multiple models and returns as a data frame.
}
\details{
This function is used within \code{residual_plots} and is not likely to be of
much use to end users.
}
\examples{
# Simulate data

N = 50

m = rep(1:4, each=N)

x1 = rnorm(n=N*4)
x2 = rnorm(n=N*4)
x3 = rnorm(n=N*4)

y = x1 + x2 + x3 + rnorm(n=N*4)

dat = data.frame(m, x1, x2, x3, y)

m1 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==1)
m2 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==2)
m3 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==3)
m4 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==4)

mList = list(m1, m2, m3, m4)

get_resid_df(model_list = mList)

}
