% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_levene.R
\name{residual_levene}
\alias{residual_levene}
\title{Compare regression residual standard deviation across multiple models.}
\usage{
residual_levene(model_list)
}
\arguments{
\item{model_list}{A list of regression models.}
}
\value{
Numeric vector of Levene test results. This includes an F-statistic,
 numerator  degrees of freedom, denominator degrees of freedom, and the
 associated p-value.
}
\description{
Compare regression residual standard deviation across multiple \code{lm}
models using Levene's test.
}
\details{
Levene's test is often suggested as a diagnostic check for homoskedasticity in
ANOVA analyses. This function works for linear regression (\code{\link{lm}})
only.
}
\examples{
N = 500

m = rep(1:4, each=N)

x1 = rnorm(n=N*4)
x2 = rnorm(n=N*4)
x3 = rnorm(n=N*4)

y = x1 + x2 + x3 + rnorm(n=N*4)

dat = data.frame(m, x1, x2, x3, y)

m1 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==1)
m2 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==2)
m3 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==3)
m4 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==4)

mList = list(m1, m2, m3, m4)

residual_levene(model_list = mList)

}
