% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_plots}
\alias{residual_plots}
\alias{residual_histogram}
\alias{residual_boxplot}
\title{Plot residuals.}
\usage{
residual_plots(model_list, model_names = NULL, thm = NULL)

residual_histogram(model_list, model_names = NULL)

residual_boxplot(model_list, model_names = NULL)
}
\arguments{
\item{model_list}{A list of regression models.}

\item{model_names}{A list of names for the regression models (default is
\code{NULL}).}

\item{thm}{A ggplot2 theme for the output plots.}
}
\value{
A residual plot for inputted regression models.
}
\description{
Plots residuals of multiple models as a histogram and a boxplot.
}
\details{
\code{residual_plots} plots histograms and boxplots in a two row, single
column format. \code{residual_histogram} and \code{residual_boxplots} plot the
component pieces. If \code{model_names = NULL}, the default, models are
numbered sequentially in the order they appear in \code{model_list} (Model 1,
Model 2, Model 3, etc.).
}
\examples{
 # Simulate data

 N = 500

 m = rep(1:4, each=N)

 x1 = rnorm(n=N*4)
 x2 = rnorm(n=N*4)
 x3 = rnorm(n=N*4)

 y = x1 + x2 + x3 + rnorm(n=N*4)

 dat = data.frame(m, x1, x2, x3, y)

 m1 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==1)
 m2 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==2)
 m3 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==3)
 m4 = lm(y ~ x1 + x2 + x3, data=dat, subset=m==4)

 mList = list(m1, m2, m3, m4)

 residual_plots(model_list = mList, thm=theme_minimal())

}
